test(morpho-sdk): colocate vault redemption actions
