+++ packages/morpho-sdk/src/actions/vaultV2/inKindRedeem.test.ts
@@ -1,20 +1,11 @@
 import { MarketParams, registerCustomAddresses } from "@morpho-org/blue-sdk";
-import {
-  decodeFunctionData,
-  maxUint256,
-  serializeSignature,
-  zeroHash,
-} from "viem";
+import { decodeFunctionData, maxUint256, serializeSignature } from "viem";
 import { describe, expect, test } from "vitest";
-import { vaultExitBundlesV1Abi } from "../abis.js";
-import {
-  EmptyMarketParamsListError,
-  type PermitRequirementSignature,
-} from "../types/index.js";
-import { vaultV1InKindRedeem } from "./vaultV1/inKindRedeem.js";
-import { vaultV2InKindRedeem } from "./vaultV2/inKindRedeem.js";
+import { vaultExitBundlesV1Abi } from "../../abis.js";
+import type { PermitRequirementSignature } from "../../types/index.js";
+import { vaultV2InKindRedeem } from "./inKindRedeem.js";
 
-const chainId = 31_337;
+const chainId = 31_338;
 const blue = "0x0000000000000000000000000000000000000001" as const;
 const vault = "0x0000000000000000000000000000000000000002" as const;
 const adapter = "0x0000000000000000000000000000000000000003" as const;
@@ -49,7 +40,7 @@ registerCustomAddresses({
   },
 });
 
-const permit = (): PermitRequirementSignature => ({
+const permit: PermitRequirementSignature = {
   args: {
     owner: userAddress,
     nonce: 7n,
@@ -66,70 +57,10 @@ const permit = (): PermitRequirementSignature => ({
       deadline: 1_900_000_000n,
     },
   },
-});
+};
 
-describe("vault in-kind redeem actions", () => {
-  test("default: Vault V1 encodes the empty-permit sentinel", () => {
-    const tx = vaultV1InKindRedeem({
-      vault: { chainId, address: vault },
-      args: {
-        amount: 100n,
-        marketParamsList: [marketParams],
-        userAddress,
-        deadline: 1_900_000_000n,
-      },
-    });
-    const decoded = decodeFunctionData({
-      abi: vaultExitBundlesV1Abi,
-      data: tx.data,
-    });
-
-    expect(tx.to).toBe(vaultExitBundlesV1);
-    expect(tx.value).toBe(0n);
-    expect(tx.action.type).toBe("vaultV1InKindRedeem");
-    expect(decoded.functionName).toBe(
-      "vaultExitBundlesV1InKindRedemptionVaultV1",
-    );
-    expect(decoded.args?.[3]).toEqual({
-      value: maxUint256,
-      nonce: 0n,
-      deadline: 1_900_000_000n,
-      v: 0,
-      r: zeroHash,
-      s: zeroHash,
-    });
-    expect(decoded).toMatchInlineSnapshot(`
-      {
-        "args": [
-          "0x0000000000000000000000000000000000000002",
-          [
-            {
-              "collateralToken": "0x0000000000000000000000000000000000000007",
-              "irm": "0x0000000000000000000000000000000000000009",
-              "lltv": 860000000000000000n,
-              "loanToken": "0x0000000000000000000000000000000000000006",
-              "oracle": "0x0000000000000000000000000000000000000008",
-            },
-          ],
-          100n,
-          {
-            "deadline": 1900000000n,
-            "nonce": 0n,
-            "r": "0x0000000000000000000000000000000000000000000000000000000000000000",
-            "s": "0x0000000000000000000000000000000000000000000000000000000000000000",
-            "v": 0,
-            "value": 115792089237316195423570985008687907853269984665640564039457584007913129639935n,
-          },
-          1900000000n,
-        ],
-        "functionName": "vaultExitBundlesV1InKindRedemptionVaultV1",
-      }
-    `);
-    expect(Object.isFrozen(tx)).toBe(true);
-    expect(Object.isFrozen(marketParams)).toBe(false);
-  });
-
-  test("default: Vault V2 encodes a signed max-share permit", () => {
+describe("vaultV2InKindRedeem", () => {
+  test("default", () => {
     const tx = vaultV2InKindRedeem({
       vault: { chainId, address: vault },
       args: {
@@ -138,7 +69,7 @@ describe("vault in-kind redeem actions", () => {
         marketParamsList: [marketParams],
         userAddress,
         deadline: 1_900_000_000n,
-        requirementSignature: permit(),
+        requirementSignature: permit,
       },
     });
     const decoded = decodeFunctionData({
@@ -184,19 +115,5 @@ describe("vault in-kind redeem actions", () => {
         "functionName": "vaultExitBundlesV1InKindRedemptionVaultV2",
       }
     `);
-  });
-
-  test("error: EmptyMarketParamsListError", () => {
-    expect(() =>
-      vaultV1InKindRedeem({
-        vault: { chainId, address: vault },
-        args: {
-          amount: 100n,
-          marketParamsList: [],
-          userAddress,
-          deadline: 1_900_000_000n,
-        },
-      }),
-    ).toThrow(EmptyMarketParamsListError);
   });
 });
